requirements.txt
python-docx>=0.8.11
python-pptx>=0.6.21
openai>=1.0.0
python-dotenv>=1.0.0
typer[all]>=0.9.0 